+++ src/app/cli/src/app.rs
@@ -105,6 +105,12 @@ pub async fn run(workspace_layout: WorkspaceLayout, args: cli::Cli) -> Result<()
     } else {
         TenancyConfig::SingleTenant
     };
+    let is_in_workspace = workspace_svc.is_in_workspace();
+    let workspace_status = match (maybe_init_command.is_some(), is_in_workspace) {
+        (false, false) => WorkspaceStatus::NoWorkspace,
+        (true, false) => WorkspaceStatus::AboutToBeCreated(tenancy_config),
+        (_, true) => WorkspaceStatus::Created(tenancy_config),
+    };
 
     let config = load_config(&workspace_layout);
     let current_account = AccountService::current_account_indication(
@@ -115,9 +121,7 @@ pub async fn run(workspace_layout: WorkspaceLayout, args: cli::Cli) -> Result<()
 
     prepare_run_dir(&workspace_layout.run_info_dir);
 
-    let is_init_command = maybe_init_command.is_some();
-    let app_database_config =
-        get_app_database_config(&workspace_layout, &config, tenancy_config, is_init_command);
+    let app_database_config = get_app_database_config(&workspace_layout, &config, workspace_status);
     let (database_config, maybe_temp_database_path) = app_database_config.into_inner();
     let maybe_db_connection_settings = database_config
         .as_ref()
@@ -134,13 +138,7 @@ pub async fn run(workspace_layout: WorkspaceLayout, args: cli::Cli) -> Result<()
             is_e2e_testing,
         );
 
-        // TODO: Use SQLite database in single-tenant
-        //       https://github.com/kamu-data/kamu-cli/issues/981
-        //
-        //       After implementing this ticket, we need to use "is_init_command"
-        //       not "init_multi_tenant_workspace" here
-        let is_indexing_needed = init_multi_tenant_workspace || workspace_svc.is_in_workspace();
-        if is_indexing_needed {
+        if workspace_status.is_indexing_needed() {
             base_catalog_builder.add::<kamu_datasets_services::DatasetEntryIndexer>();
             base_catalog_builder.add::<kamu_datasets_services::DependencyGraphIndexer>();
             base_catalog_builder.add::<kamu_auth_rebac_services::RebacIndexer>();
@@ -176,7 +174,7 @@ pub async fn run(workspace_layout: WorkspaceLayout, args: cli::Cli) -> Result<()
         register_config_in_catalog(
             &config,
             &mut base_catalog_builder,
-            tenancy_config,
+            workspace_status,
             is_e2e_testing,
         );
 
@@ -223,7 +221,7 @@ pub async fn run(workspace_layout: WorkspaceLayout, args: cli::Cli) -> Result<()
 
     let is_workspace_upgrade_needed = workspace_svc.is_upgrade_needed()?;
 
-    if workspace_svc.is_in_workspace() && !is_workspace_upgrade_needed {
+    if is_in_workspace && !is_workspace_upgrade_needed {
         // Evict cache
         cli_catalog.get_one::<GcService>()?.evict_cache()?;
     }
@@ -242,7 +240,7 @@ pub async fn run(workspace_layout: WorkspaceLayout, args: cli::Cli) -> Result<()
             let mut command =
                 cli_commands::get_command(work_catalog, &maybe_transactional_cli_catalog, args)?;
 
-            if command.needs_workspace() && !workspace_svc.is_in_workspace() {
+            if command.needs_workspace() && !is_in_workspace {
                 Err(CLIError::usage_error_from(NotInWorkspace))?;
             }
             if command.needs_workspace() && is_workspace_upgrade_needed {
@@ -496,6 +494,7 @@ pub fn configure_base_catalog(
 
     b.add::<DatabaseTransactionRunner>();
 
+    b.add::<kamu_auth_rebac_services::RebacDatasetLifecycleMessageConsumer>();
     b.add::<kamu_auth_rebac_services::RebacServiceImpl>();
     b.add_value(kamu_auth_rebac_services::DefaultAccountProperties { is_admin: false });
     b.add_value(kamu_auth_rebac_services::DefaultDatasetProperties {
@@ -512,10 +511,6 @@ pub fn configure_base_catalog(
             .unwrap(),
     );
     b.add::<kamu_adapter_flight_sql::KamuFlightSqlService>();
-
-    if tenancy_config == TenancyConfig::MultiTenant {
-        b.add::<kamu_auth_rebac_services::MultiTenantRebacDatasetLifecycleMessageConsumer>();
-    }
 
     b.add::<kamu_datasets_services::DatasetEntryServiceImpl>();
     b.add::<kamu_datasets_services::DependencyGraphServiceImpl>();
@@ -633,7 +628,7 @@ fn load_config(workspace_layout: &WorkspaceLayout) -> config::CLIConfig {
 pub fn register_config_in_catalog(
     config: &config::CLIConfig,
     catalog_builder: &mut CatalogBuilder,
-    tenancy_config: TenancyConfig,
+    workspace_status: WorkspaceStatus,
     is_e2e_testing: bool,
 ) {
     let network_ns = config.engine.as_ref().unwrap().network_ns.unwrap();
@@ -758,37 +753,43 @@ pub fn register_config_in_catalog(
     catalog_builder.add_value(flight_sql_conf.to_session_auth_config());
     catalog_builder.add_value(flight_sql_conf.to_session_caching_config());
 
-    if tenancy_config == TenancyConfig::MultiTenant {
-        let mut implicit_user_config = PredefinedAccountsConfig::new();
-        implicit_user_config.predefined.push(
-            AccountConfig::test_config_from_name(odf::AccountName::new_unchecked(
-                AccountService::default_account_name(TenancyConfig::MultiTenant).as_str(),
-            ))
-            .set_display_name(AccountService::default_user_name(
-                TenancyConfig::MultiTenant,
-            )),
-        );
-
-        if is_e2e_testing {
-            let e2e_user_config =
-                AccountConfig::test_config_from_name(odf::AccountName::new_unchecked("e2e-user"));
-
-            implicit_user_config.predefined.push(e2e_user_config);
-        }
-
-        use merge::Merge;
-        let mut user_config = config.users.clone().unwrap();
-        user_config.merge(implicit_user_config);
-        catalog_builder.add_value(user_config);
-    } else {
-        if let Some(users) = &config.users {
-            assert!(
-                users.predefined.is_empty(),
-                "There cannot be predefined users in a single-tenant workspace"
+    if let Some(tenancy_config) = workspace_status.into_tenancy_config() {
+        if tenancy_config == TenancyConfig::MultiTenant {
+            let mut implicit_user_config = PredefinedAccountsConfig::new();
+            implicit_user_config.predefined.push(
+                AccountConfig::test_config_from_name(odf::AccountName::new_unchecked(
+                    AccountService::default_account_name(TenancyConfig::MultiTenant).as_str(),
+                ))
+                .set_display_name(AccountService::default_user_name(
+                    TenancyConfig::MultiTenant,
+                )),
             );
-        }
 
-        catalog_builder.add_value(PredefinedAccountsConfig::single_tenant());
+            if is_e2e_testing {
+                let e2e_user_config = AccountConfig::test_config_from_name(
+                    odf::AccountName::new_unchecked("e2e-user"),
+                );
+
+                implicit_user_config.predefined.push(e2e_user_config);
+            }
+
+            use merge::Merge;
+            let mut user_config = config.users.clone().unwrap();
+            user_config.merge(implicit_user_config);
+            catalog_builder.add_value(user_config);
+        } else {
+            if let Some(users) = &config.users {
+                assert!(
+                    users.predefined.is_empty(),
+                    "There cannot be predefined users in a single-tenant workspace"
+                );
+            }
+
+            catalog_builder.add_value(PredefinedAccountsConfig::single_tenant());
+        }
+    } else {
+        // No workspace
+        catalog_builder.add_value(PredefinedAccountsConfig::new());
     }
 
     let uploads_config = config.uploads.as_ref().unwrap();
@@ -834,6 +835,30 @@ pub fn register_config_in_catalog(
         Duration::seconds(outbox_config.awaiting_step_secs.unwrap()),
         outbox_config.batch_size.unwrap(),
     ));
+}
+
+#[derive(Debug, Copy, Clone, Eq, PartialEq)]
+pub enum WorkspaceStatus {
+    NoWorkspace,
+    AboutToBeCreated(TenancyConfig),
+    Created(TenancyConfig),
+}
+
+impl WorkspaceStatus {
+    fn into_tenancy_config(self) -> Option<TenancyConfig> {
+        match self {
+            WorkspaceStatus::NoWorkspace => None,
+            WorkspaceStatus::AboutToBeCreated(tenancy_config)
+            | WorkspaceStatus::Created(tenancy_config) => Some(tenancy_config),
+        }
+    }
+
+    fn is_indexing_needed(self) -> bool {
+        match self {
+            WorkspaceStatus::NoWorkspace => false,
+            WorkspaceStatus::AboutToBeCreated(_) | WorkspaceStatus::Created(_) => true,
+        }
+    }
 }
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////