+++ src/app/cli/tests/tests/test_di_graph.rs
@@ -11,7 +11,7 @@ use kamu::domain::{ServerUrlConfig, TenancyConfig};
 use kamu_accounts::{CurrentAccountSubject, JwtAuthenticationConfig};
 use kamu_adapter_http::AccessToken;
 use kamu_adapter_oauth::GithubAuthenticationConfig;
-use kamu_cli::{self, Interact, OutputConfig, WorkspaceLayout};
+use kamu_cli::{self, Interact, OutputConfig, WorkspaceLayout, WorkspaceStatus};
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
@@ -56,7 +56,7 @@ fn test_di_cli_graph_validates(tenancy_config: TenancyConfig) {
     kamu_cli::register_config_in_catalog(
         &kamu_cli::config::CLIConfig::default(),
         &mut base_catalog_builder,
-        tenancy_config,
+        WorkspaceStatus::Created(tenancy_config),
         false,
     );
     base_catalog_builder.add_value(Interact::new(false, false));
@@ -90,7 +90,12 @@ fn test_di_server_graph_validates(tenancy_config: TenancyConfig) {
 
     let config = kamu_cli::config::CLIConfig::default();
 
-    kamu_cli::register_config_in_catalog(&config, &mut base_catalog_builder, tenancy_config, false);
+    kamu_cli::register_config_in_catalog(
+        &config,
+        &mut base_catalog_builder,
+        WorkspaceStatus::Created(tenancy_config),
+        false,
+    );
     let base_catalog = base_catalog_builder.build();
 
     let mut cli_catalog_builder = kamu_cli::configure_server_catalog(&base_catalog);