+++ src/e2e/app/cli/repo-tests/src/commands/test_init_command.rs
@@ -7,64 +7,128 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0.
 
-use kamu_cli::{DEFAULT_MULTI_TENANT_SQLITE_DATABASE_NAME, KAMU_WORKSPACE_DIR_NAME};
+use kamu::domain::TenancyConfig;
+use kamu_cli::{DEFAULT_WORKSPACE_SQLITE_DATABASE_NAME, KAMU_WORKSPACE_DIR_NAME};
 use kamu_cli_puppet::extensions::KamuCliPuppetExt;
 use kamu_cli_puppet::KamuCliPuppet;
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+// test_init_creates_sqlite_database
+////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
-pub async fn test_init_multi_tenant_creates_sqlite_database(mut kamu: KamuCliPuppet) {
+pub async fn test_init_creates_sqlite_database_st(kamu: KamuCliPuppet) {
+    test_init_creates_sqlite_database(kamu, TenancyConfig::SingleTenant).await;
+}
+
+////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+
+pub async fn test_init_creates_sqlite_database_mt(kamu: KamuCliPuppet) {
+    test_init_creates_sqlite_database(kamu, TenancyConfig::MultiTenant).await;
+}
+
+////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+// test_init_with_exists_ok_flag_creates_sqlite_database
+////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+
+pub async fn test_init_with_exists_ok_flag_creates_sqlite_database_st(kamu: KamuCliPuppet) {
+    test_init_with_exists_ok_flag_creates_sqlite_database(kamu, TenancyConfig::SingleTenant).await;
+}
+
+////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+
+pub async fn test_init_with_exists_ok_flag_creates_sqlite_database_mt(kamu: KamuCliPuppet) {
+    test_init_with_exists_ok_flag_creates_sqlite_database(kamu, TenancyConfig::MultiTenant).await;
+}
+
+////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+// test_init_exist_ok
+////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+
+pub async fn test_init_exist_ok_st(kamu: KamuCliPuppet) {
+    test_init_exist_ok(kamu, TenancyConfig::SingleTenant).await;
+}
+
+////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+
+pub async fn test_init_exist_ok_mt(kamu: KamuCliPuppet) {
+    test_init_exist_ok(kamu, TenancyConfig::MultiTenant).await;
+}
+
+////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+// test_init_in_an_existing_workspace
+////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+
+pub async fn test_init_in_an_existing_workspace_st(kamu: KamuCliPuppet) {
+    test_init_in_an_existing_workspace(kamu, TenancyConfig::SingleTenant).await;
+}
+
+////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+
+pub async fn test_init_in_an_existing_workspace_mt(kamu: KamuCliPuppet) {
+    test_init_in_an_existing_workspace(kamu, TenancyConfig::MultiTenant).await;
+}
+
+////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+// Implementations
+////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+
+async fn test_init_creates_sqlite_database(mut kamu: KamuCliPuppet, tenant_config: TenancyConfig) {
     kamu.set_workspace_path_in_tmp_dir();
 
-    kamu.execute(["init", "--multi-tenant"]).await.success();
+    kamu.execute(match tenant_config {
+        TenancyConfig::SingleTenant => vec!["init"],
+        TenancyConfig::MultiTenant => vec!["init", "--multi-tenant"],
+    })
+    .await
+    .success();
 
     let expected_database_path = kamu
         .workspace_path()
         .join(KAMU_WORKSPACE_DIR_NAME)
-        .join(DEFAULT_MULTI_TENANT_SQLITE_DATABASE_NAME);
+        .join(DEFAULT_WORKSPACE_SQLITE_DATABASE_NAME);
 
     assert!(expected_database_path.exists());
 }
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
-pub async fn test_init_multi_tenant_with_exists_ok_flag_creates_sqlite_database(
+async fn test_init_with_exists_ok_flag_creates_sqlite_database(
     mut kamu: KamuCliPuppet,
+    tenant_config: TenancyConfig,
 ) {
     kamu.set_workspace_path_in_tmp_dir();
 
-    kamu.execute(["init", "--multi-tenant", "--exists-ok"])
-        .await
-        .success();
+    kamu.execute(match tenant_config {
+        TenancyConfig::SingleTenant => vec!["init", "--exists-ok"],
+        TenancyConfig::MultiTenant => vec!["init", "--multi-tenant", "--exists-ok"],
+    })
+    .await
+    .success();
 
     let expected_database_path = kamu
         .workspace_path()
         .join(KAMU_WORKSPACE_DIR_NAME)
-        .join(DEFAULT_MULTI_TENANT_SQLITE_DATABASE_NAME);
+        .join(DEFAULT_WORKSPACE_SQLITE_DATABASE_NAME);
 
     assert!(expected_database_path.exists());
 }
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
-pub async fn test_init_exist_ok_st(mut kamu: KamuCliPuppet) {
+async fn test_init_exist_ok(mut kamu: KamuCliPuppet, tenant_config: TenancyConfig) {
     kamu.set_workspace_path_in_tmp_dir();
 
-    kamu.execute(["init"]).await.success();
-    kamu.execute(["init", "--exists-ok"]).await.success();
-}
-
-////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-
-pub async fn test_init_exist_ok_mt(mut kamu: KamuCliPuppet) {
-    kamu.set_workspace_path_in_tmp_dir();
-
-    kamu.execute(["init", "--multi-tenant"]).await.success();
+    kamu.execute(match tenant_config {
+        TenancyConfig::SingleTenant => vec!["init"],
+        TenancyConfig::MultiTenant => vec!["init", "--multi-tenant"],
+    })
+    .await
+    .success();
 
     let expected_database_path = kamu
         .workspace_path()
         .join(KAMU_WORKSPACE_DIR_NAME)
-        .join(DEFAULT_MULTI_TENANT_SQLITE_DATABASE_NAME);
+        .join(DEFAULT_WORKSPACE_SQLITE_DATABASE_NAME);
 
     let modified_old = expected_database_path
         .metadata()
@@ -72,9 +136,12 @@ pub async fn test_init_exist_ok_mt(mut kamu: KamuCliPuppet) {
         .modified()
         .unwrap();
 
-    kamu.execute(["init", "--multi-tenant", "--exists-ok"])
-        .await
-        .success();
+    kamu.execute(match tenant_config {
+        TenancyConfig::SingleTenant => vec!["init", "--exists-ok"],
+        TenancyConfig::MultiTenant => vec!["init", "--multi-tenant", "--exists-ok"],
+    })
+    .await
+    .success();
 
     let modified_new = expected_database_path
         .metadata()
@@ -88,14 +155,26 @@ pub async fn test_init_exist_ok_mt(mut kamu: KamuCliPuppet) {
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
-pub async fn test_init_in_an_existing_workspace(mut kamu: KamuCliPuppet) {
+async fn test_init_in_an_existing_workspace(mut kamu: KamuCliPuppet, tenant_config: TenancyConfig) {
     kamu.set_workspace_path_in_tmp_dir();
 
-    kamu.assert_success_command_execution(["init"], None, Some(["Initialized an empty workspace"]))
-        .await;
+    let command = match tenant_config {
+        TenancyConfig::SingleTenant => vec!["init"],
+        TenancyConfig::MultiTenant => vec!["init", "--multi-tenant"],
+    };
+
+    kamu.assert_success_command_execution(
+        &command,
+        None,
+        Some(match tenant_config {
+            TenancyConfig::SingleTenant => vec!["Initialized an empty workspace"],
+            TenancyConfig::MultiTenant => vec!["Initialized an empty multi-tenant workspace"],
+        }),
+    )
+    .await;
 
     kamu.assert_failure_command_execution(
-        ["init"],
+        &command,
         None,
         Some(["Error: Directory is already a kamu workspace"]),
     )