+++ src/utils/repo-tools/src/release.rs
@@ -116,7 +116,7 @@ fn update_license(
 ) {
     let text = std::fs::read_to_string(license_path).expect("Could not read the license file");
     let new_text = update_license_text(&text, current_version, new_version, current_date);
-    assert_ne!(text, new_text);
+    pretty_assertions::assert_ne!(text, new_text);
     std::fs::write(license_path, new_text).expect("Failed to write to license file");
 }
 
@@ -130,7 +130,7 @@ fn update_openapi_schema(path: &Path, new_version: &Version) {
         })
         .to_string();
 
-    assert_ne!(text, new_text);
+    pretty_assertions::assert_ne!(text, new_text);
     std::fs::write(path, new_text).expect("Failed to write to schema file");
 }
 
@@ -162,14 +162,16 @@ fn update_license_text(
 fn update_changelog(path: &Path, new_version: &Version, current_date: NaiveDate) {
     let text = std::fs::read_to_string(path).expect("Could not read the changelog file");
 
-    let re = regex::Regex::new(r#"## +\[?Unreleased\]? *"#).unwrap();
+    let re = regex::Regex::new(r#"(?m)^## +\[?Unreleased\]? *"#).unwrap();
     let new_text = re
         .replace(&text, |_: &Captures| {
             format!("## [{new_version}] - {current_date}")
         })
         .to_string();
 
-    assert_ne!(text, new_text, "Unreleased changes section not found");
+    // The usage of assert!() instead of assert_ne!() is intentional:
+    // we don't want to output both [large] files
+    assert!({ text != new_text }, "Unreleased changes section not found");
 
     std::fs::write(path, new_text).expect("Failed to write to changelog file");
 }
@@ -177,11 +179,11 @@ fn update_changelog(path: &Path, new_version: &Version, current_date: NaiveDate)
 fn update_web_ui_version_for_release(path: &Path) {
     fn get_latest_version() -> Version {
         // To avoid requiring the use of GITHUB_TOKEN here, we use a little trick
-        const DESKTOP_LINE_USER_AGENT: &str =
+        const DESKTOP_LIKE_USER_AGENT: &str =
             "Mozilla/5.0 (X11; Linux x86_64; rv:134.0) Gecko/20100101 Firefox/134.0";
 
         let client = reqwest::blocking::Client::builder()
-            .user_agent(DESKTOP_LINE_USER_AGENT)
+            .user_agent(DESKTOP_LIKE_USER_AGENT)
             .build()
             .unwrap();
 
@@ -264,7 +266,7 @@ mod tests {
             NaiveDate::from_str("2021-09-01").unwrap(),
         );
 
-        assert_eq!(
+        pretty_assertions::assert_eq!(
             new_text,
             indoc::indoc!(
                 r#"
@@ -303,7 +305,7 @@ mod tests {
             NaiveDate::from_str("2021-09-01").unwrap(),
         );
 
-        assert_eq!(
+        pretty_assertions::assert_eq!(
             new_text,
             indoc::indoc!(
                 r#"